Basic code for altered minimizing player (not yet functional)
Added a "maximizing" parameter to bound which reflects the active
player.  Added global vars BEST_PENALTY and SECOND_BONUS, which
determine the margin for considering a second-best move.

Currently, minimizing bound still returns best. This results in cutoffs
due to secondbest being scored lower than they should be, and not
incentivizing the right behavior. I'm having it return best instead of
val for now anyway because best is the value of an actual move and it
won't cause fatal problems.
I may have to implement code that stores the first/second best moves
along with their scores when running through moves and increase the
number of killer moves to 2. This would ensure that moves are examined
in the right order during a second run-through and that the value of a
node stays constant within the same search.